[mycpp/runtime] Move string length out of object header
Making room for a bigger 30-bit object ID

To avoid bloating metrics, we're using the unused hash value field for
now.

Remove the #ifdef for Cheney heap.  It seems unlikely we'll use it, so
let's keep the code simple.

---

Also remove bad % usage which caused memory unsafety in C++ with
StrFormat().  I wonder if we can add some type checking to mycpp.