Add usage info to test/spec.sh
Make test/spec.sh slightly friendlier/discoverable by listing available
functions/subcommands when invoked with -h/--help or no args.

This uses a simplistic method to list the functions which has problems
(main is listed, for example) but it's a good start and goes most of
the way to making it more usable for new contributors.