[core optimize] SignalSafe::ReuseEmptyList() to avoid many allocations
We TakePendingSignals() in the main interpreter loop, which shows up on
benchmarks/compute/fib.sh.

This change lets us avoid ~3K of ~150K allocations on that benchmark, or
about 2%.  These allocations also wouldn't fit in the proposed pool
allocator for small objects, so they would result in a malloc() and
free(), which is slow.