[osh-language] Implement new eval_unsafe_arith on unset, printf, ${!ref}
This is related to the fix for #1450.  It relaxes eval_unsafe_arith in
the remaining places.

One more test in spec/nix-idioms now fails.  It passed for the wrong
reason -- it because 'foo[@]' wasn't a valid var ref, not because we
implemented the weird and arguably buggy "empty array + var ref" rule
that bash has.

Also:

- Remove eval_unsafe_arith from almost all spec tests.
- Update docs.