[pyext/line_input] Remove #ifdef in method tables
build/cpython-defs.sh runs the preprocessor on these method tables, and
it doesn't have the #defines (which are always set in the file).

The new spec tests caught that 'history -c' called clear_history(),
which wasn't being built into the binary!

Also add 'test/spec-cpp.sh ysh-all' to the release.