[core/shell] Use $HOME before getpwuid() to get home dir
This is more reliable, and addresses the condition in #1578.  The login
program is supposed to set $HOME.

In the unlikely case that both methods fail, it's a fatal error.

Verify it with test/ltrace.sh, which is hooked up to CI

Also add 'osh' or 'ysh' prefix to many log messages