[mycpp] fix Callable -> func pointer translation
There's a bug in the syntax emitted for function pointers (not seemingly used
until now). In all contexts, function pointer are declared like variables and
given the same name, `f`. This 1) makes it impossible to have more than one
function pointer on the stack, and 2) makes it impossible to use function
pointers as template parameters. This patch tweaks GetCType() to return a
decomposed representation of the type for function pointers. This way callers
can reassemble the parts in a context-appropriate manner.