[eggex breaking] Disallow $x ${x} "$x ${x}" in eggex
Instead allow splicing string values only with @var_name.

The @ syntax is meant as a hint that it does ESCAPING rather than raw
substitution.

This is a BIG SIMPLIFICATION of eggex.  The extra $x ${x} stuff wasn't
documented in doc/eggex.md anyway.

This is issue #1639.  I noticed the design problem on a video call with
Melvin.