[core] create jobs for procs when sending to bg... and add basic job specs to `fg` builtin
Right now, only jobs started in the background get added to the jobs
list. If some, say, starts `vim` in the foreground and sends it to the
background with `Ctrl-z` the output of `jobs` will be empty. `fg` will
still bring it back to the foreground as long as it's considered the
last stopped process, but sending another process to the background will
prevent the user from ever bringing the original process back to the
foreground.

To make managing multiple background processes easier this commit adds
processes stopped with `Ctrl-z` to the jobs list and adds support for
the `%<digits>` flavor of job spec to the `fg` builtin.