[core] only take back tty on stop if necessary
When the main shell wakes up from wait() with WIFSTOPPED, it
unconditionally takes back the TTY from whichever process was just
stopped. This can cause problems when running background jobs, though.

Consider the following script:

  $ cat &
  $ cat

The first cat is forked and placed into its own process group. The shell
remains in the foreground. This means cat will immediately be stopped
with SIGTTIN when it tries to read from stdin. The main shell doesn't
know this yet, though! It's blocked waiting for input from readline. The
second cat is forked and placed into its own group, just like the first
one; and since it's running in the foreground, the main shell gives it
control of the TTY. After handing over the TTY the shell finally gets
notified about the first cat being stopped (as expected) via wait() and
does some post-stop bookkeeping, including taking back the TTY! This
causes the second cat to immediately get stopped by SIGTTIN, confusing
the user.

To avoid this awkward behavior, this commit updates the post-stop
bookkeeping to only take back the terminal from processes that were
originally running in the foreground.