[core] print notifications when bg jobs complete
To support notifying users about the completion of background jobs, this
commit updates the main loop for the interactive shell to pass the
WNOHANG to `WaitPid()`, adds state to `Job` about background/foreground
positioning. It also removes the `WhenDone()` method from `JobList` to
make reading the job completion codepath easier.