[pea] Nice test of parallel parsing
We can parse in under 300ms by using 8 cores.

It would nice to parse, type check, and emit C++ for the whole codebase
in under a second.

Or just parsing and type checking in one second would be an improvement.

I think that should be easily doable, maybe even in Python.  Maybe we
can use fast 'marshal' serialization or something.

MyPy (without mycpp speedups) is taking 4-5 seconds to parse and type
check.

With the mycpp speedups, it's possible it will get down to 1 second, not
sure.  Could investigate that build.