[errors] Remove SetTokenForLine() in command.If
The child command nodes will take care of it.

I have a suspicion that this will show up in benchmarks2/uftrace, in a
very minor way.  But even if it doesn't, the code is more consistent
now.

We can still improve errors messages through SetLocationForExpr() and
attaching loc_t to each specific error.

Part of #1707

Minor: PlaceMutation keyword isn't optional