[testdata/completion] demo: bash COMPREPLY is a shell string
It must be shell-escaped, like

   word\ with\ spaces
   can\'t
   $'file \n with newline'

'printf %q' is the right thing.  It works with space, single quote,
newline, etc.

Another consequence of COMPREPLY being a shell string is that the
trailing space is a no-op.  It's a space in the shell command, not a
space in the arg being completed.