[ysh] Tighten up float syntax
Not allowed

    42.  # Must be 42.0

This makes it easier to use this syntax for ranges

    = 1..5

It's also consistent with this being illegal

    .333  # must be 0.333

Remove the LEXER_REFINEMENTS thing.  Not sure why I added that, since we
can just remove _ everywhere.

Oh I think I wanted that to avoid parsing the float twice.  Now we just
replace _ with '' and pass to float().

float() uses libc strtod() under the hood, which seems reasonable.  I
think I had added that for some kind of LST invariant or separate
parsing, but strtod() is fine.