[ysh breaking] s->trim(), s->startsWith()
camelCase: startswith() -> startsWith()

See Zulip for rationale on s->trim() over s->strip().

We're not implementing all of Python's API, and there are a number of
cosmetic reasons for trim().

Settled on this API:

   trim()
   trimLeft()
   trimRight()

   trimPrefix()
   trimSuffix()

Though to be honest I wonder if the "speed bump" with Python will be
annoying.  It's not that hard for us to accept 'startswith' as well.

Although I want one way to do it, when methods in Python/JS differ by
CASE ONLY, it could be an argument for having both names.