[ysh-language] Implement function threading / chaining with =>
The expression

    obj => f

attempts to create a value.BoundFunc.

First, it's a method lookup on the type of 'obj'.  (Right now we only
have built-in methods, no user-defined methods.)

If a specific method isn't found, then looks up a variable named
'f'.

If there's no such variable, then it we get a normal 'Undefined
variable' error.

If it's found and it's a BuiltinFunc or (user-defined) Func, we create a
BoundFunc.

---

So this makes FREE funtions chain like methods.  The example from
spec/ysh-methods.test.sh shows the benefit.

If dictfunc() returns a dict with keys K1 and K2, then

Old way:

    $ echo $[list(dictfunc()) => join('/') => upper()]
    => K1/K2

New way that's nicer and more consistent

    $ echo $[dictfunc() => list() => join('/') => upper()]
    => K1/K2

TODO: Write this down in doc/ref.