[eggex breaking] Change capture syntax
Old syntax:

    var pat = / <d+> /                   # positional

    var pat = / <d+ : month> /           # named

New Syntax

    var pat = / <capture d+> /           # positional

    var pat = / <capture d+ as month> /  # named

This is mainly an aesthetic change to a style more like Python.

But it also allow types, which can be used for conversions:

    var pat = / <capture d+ as month: Int> /

Also remove obsolete Tea keywords, and add 'capture' 'as'.  Eggexes are
lexed as normal expressions, so they're keywords in arithmetic
expressions too.

TODO:

- We're still not respecting named captures
- Int and Float type conversion.
- Test if libc even gives you multiple captures with repetition?  I
  think bash doesn't have this.