[j8] Migrate everything from QSN -> J8
This exposed a bug in J8!  We're using the Bjoern DFA incorrectly.

TODO: Fix the bug temporarily allowed in spec/serialize.test.sh.

We may have to output 4 copies of \yff, or 16 bytes in a single
J8EncodeOne() / ShellEncodeone() invocation.

The state machine may see up to 3 possibly bytes before it sees an
invalid byte.  The last byte determines if the previous 3 were valid or
invalid.