[j8] Fix UTF8_REJECT handling.
We may get UTF8_REJECT from the Bjoern DFA after 1, 2, 3, or 4 bytes.

In all cases, we now REWIND the input to the first step, output the
escape like \xff or \yff, and then return.

The caller makes another invocation.

We can't output all 4 bytes, because the last one could be the NUL
terminator!  We only output one invalid byte at a time.