[mycpp/runtime] Disable List/Str slice step support
We're apparently not using this anymore.

We rewrote the usages in expr_to_ast.py with loops.

And it's not in because I wanted a :: operator at some point, and a[::2]
conflicts.

We might want it at some point, but we have to fix this crash bug first,
in spec/ysh-slice-range.