[ysh parser] Fix crash with b'\x'
This is issue #1826.

It becomes Id.Unknown_Backslash, not Id.Unknown_Tok, and correctly gives
a syntax error.

Cleanup: remove Id.Unknown_Tok from lex_mode_e.SQ_C, which is for
bash $'' strings.

The same code in osh/word_parse.py recognizes the tokens in bash $'' and
YSH u''.  So the lexers should be consistent.