[test/errors] Move YSH cases to ysh-parse-errors
On Zulip, I noted that we have SEVEN of these kinds of files.  And we
want to run them on 3 binaries.

So yet another table-driven test harness would be nice.  We have that in
mycpp/examples, unit tests, Soil, etc.

I guess I want a flat list of tasks, and then we can arrange them in
tables after the fact.