[core/process] Remove getpgid() calls, in favor of self.pgid
One of them causes the ESRCH bug, triggered by starship and others.

Repro:

   $ bin/osh
   osh$ . demo/starship-like.sh
   osh$ echo hi | cat  # would crash, no longer does

This is issue #1822.

We still have the setpgid() failure though.  I think we will remove that
by disabling the DEBUG trap when running the last part of the pipeline.
At least when job control is on.

TODO: Also need unit tests for StartPipeline()