[mycpp] Expressing StackArray with MyPy
It seems like this will work!  We can create StackArray[T], which is
similar to List[T], but not interchangeable.

This is important because we never want to store a StackArray[T] inside
ASDL containers, which are currently always heap-allocated.

Now we have to do the code gen side, to make sure we can inspect the
type parameters and generate the right C++ code.