[ysh] Block arg has its own arg group - the third one
This addresses issue #1849.  I's a BREAKING change.

We don't want ...pos splats to be intertwined with the block arg.

We also want the signature to match the call site, e.g.

    proc p (...w; ...t; ...n; block) { echo }

and

    p @w (...t; ...n; block)

This means there is a distinction between block arguments and positional
value.Command arguments.

Eval has a positional value.Command

    eval (^(echo hi))

cd has a block argument:

    cd /tmp (; ; ^(echo hi))