add some statically defined tracepoints
This commit adds some infrastructure for adding static tracepoints to
the shell runtime and a few examples of how to use it.

These trace points can be used with tools like linux perf/ebpf and
dtrace to give us a better idea about what the shell is doing at
runtime.