mycpp: add ControlFlowGraph library
This commit adds a library for implementing a simple control flow graph.
Right now it doesn't have any meaningful statement annotations, nor is
it connected to anything in mycpp. Future commits will build on top of
this to generate datalog programs to perform analyses like the
call-graph analysis from #1963 or the alias or stack root analysis
from #1972.