[mycpp/float_test] Summarize articles and clean up code
- It's good to know that sprintf() and sscanf() can round trip all
  floats to decimal and back
- And that we can test the round tripping in ~12 minutes

But this is probably not relevant to Oils.  I think we will use
heap-allocated double instead of trying to fit a float in a tagged
pointer.

Using float is probably premature optimization, at the expense of
functionality.

Some examples of limitations:

https://randomascii.wordpress.com/2012/02/13/dont-store-that-in-a-float/

e.g. it can be natural to store time as a double of seconds.