[mycpp] Big refactoring, annotate CFG with function calls
* [mycpp] annotate CFG with function calls

This commit adds the first concrete `Fact` and uses it to annotate
statements that make function calls. You can feed this in to
`_bin/datalog/call-graph` and see a list of statements that might lead
to the GC being invoked. This output still doesn't impact the generated
C++, though.

* mycpp: add SimpleVisitor

* mycpp: add ir_pass

* add FooMember types

We are trying to put more heuristics in the first pass, so the later passes have a clean structure to work with.