[build/deps] Use a different version of typed_ast, for Fedora
The version of MyPy we're using for mycpp depends on typed_ast

However it is pinned to an older version that has this bug

https://github.com/python/typed_ast/issues/169

The compile error happens on Fedora (but not Debian or Alpine)

So we patch mypy-requirements.txt to use the new version, and it seems
to work.

TODO: Rebuild my wedge locally to test it.

---

Also add comments about potential algorithmic slowness for
value.SparseArray, which uses Dict[BigInt, str].