[mycpp/runtime] Add keys_equal() overload for mops::BigInt
Fixes hashing bug.

I also wonder if we can get rid of the are_equal() and keys_equal()
distinction.

I think one is used in List<T>, and the other in Dict<K, V>, but there's
no reason they can't be the same.

Oh one reason might be if Dicts can be keyed by mutable objects with
IDENTITY, e.g.  for use in a type checker.  But I think such lists
should not be?  Though arguably it would be consistent to allow both.