[test] Add jsontestsuite
This is a hacky way of running nst/JSONTestSuite. It's a starting point
for conversation, not something that's mergeable.

Run using `python3 jsontestsuite/run_tests.py` which
produces a `log.txt` and 2x html files in `results/`

The original Python script assumes a particular directory
structure and presence of certain files when producing results.

If one library is tested, the results only show failures and
implementation specific tests. (The original Python script wants to
compare results between libraries.) A simple way to get more output was
to run another library (e.g. Bash) to compare and one that isn't
installed (e.g. Awk, which the script assumes will be in a specific
directory) and will always fail.