[ysh/builtin] First pass of assert builtin
It accepts all of these forms:

    assert (false)
    assert [false]  # evaluates it for you

    assert (42 === f())  # eagerly evaluated, not special

    assert [42 === f()]  # evaluates it for you, prints error message

We do NOT yet quote the code.  But the error is already pretty good
without that.

We probably want to use the pretty printer to show the values wrapped,
and in color.

A diff may also be useful, but the user can also invoke 'diff'
themselves.