[osh] Make behavior of var x = $(echo command sub) more consistent
In YSH, we failed as expected.

But in OSH, we didn't have all the errexit options on.

We haven't fully specified the behavior of 'var' in OSH.  There are two
options:

- we could leave it out, just like we leave out proc and func now
- we could enable var (and setvar setglobal), proc, and func
  - this is the strategy of making ALL of YSH available in OSH, without
    every typing bin/ysh, or shopt --set ysh:all

There is a similar issue with set -o nounset - it is not set in a var
statement running under OSH.

I think the most principled thing to do is to turn on ysh:all inside
var.  And likewise in proc and func.

I may try that next.