[core] Optimize forks by moving decision into executor
We pass cmd_val.is_last_cmd to builtins, and they pass back
executor.IS_LAST_CMD

This makes this case more efficient:

    command date | wc -l

Prior to this change, we "lost" the info that 'date' is the last
command.  It's now identical to this:

    date | wc -l

On test/syscall, we are now more optimal.