[spec/ysh-builtin-eval] Use local vars in vars/pos_args binding cases
I was trying to find what I thought was a design issue, but couldn't tickle it.

Then I made a diversion into 'pp stacks_' to try print state.Mem in a
readable way.

That led to some diversions:

- shopt --unset copy_env - to make the output less cluttered,
- Fixed an ordering bug in printin options: bin/ysh -o

pp stacks_ isn't done yet, but could be a good starting point.  Right
now it uses a similar mechanism as the JSON crash dump, which is
supposed to dump the interpreter state in JSON.  (This isn't used very
much now, but is tested.)