[mycpp/dataflow] Add stack roots solver with --minimize-stack-roots (#2023)
This patch adds a stack roots solver written in souffle. It uses the facts and control flow graph emitted by control_flow_pass to compute a minimal (approximately) set of stack roots required for safe execution of the source program. The relations and rules for this program, AKA our rooting policy, are derived from the examples enumerated in the translation example added in #2030.

If mycpp is run with the --minimize-stack-roots flag, it will execute the stack roots solver and feed its output into cppgen_pass where it will be queried to determine if a StackRoot should be emitted for a local variable. When --minimize-stack-roots is set, cppgen_pass will also omit stack roots for loop index variables. Without --minimize-stack-roots there are no changes to the generated code produced by cppgen_pass.

Running mycpp with --minimize-stack-roots reduces the number of stack roots in _gen/bin/oils_for_unix.mycpp.cc by ~78%!

Before

$ grep StackRoot _gen/bin/oils_for_unix.mycpp.cc | wc -l
4744

After

$ grep StackRoot _gen/bin/oils_for_unix.mycpp.cc | wc -l
1001