[ysh breaking] Enforce that -> is always mutating method lookup
Even on builtin types.

Now we no longer have the "relaxed" semantics of -> and => being
interchangeable.

---

Idea for hiding the 'io'. name.  This is a bit tricky since doing it at
every function call is overkill.  Ideally, we want to do it only at the
proc/func boundary.

That can't be at ctx_Expr either, because we want echo
@[io.glob('*.py')] to work.