[builtin/trap] Revert part of refactoring
It inadvertently made a spec test pass: spec/builtin-trap.test.sh, case
20

I will restore the fix once it's clear why it happened!

Ah I think this is because we removed a call to:

    pyos.RegisterSignalInterest(SIGINT)

which calls

    signal.signal(SIGINT)

which affects KeyboardInterrupt.

This is related to the trap INT bug, but I'm not sure what the real fix
is.

OSH matches other shells now, besides mksh.  But I thought mksh behavior
was OK!