[osh-py] SIGINT handler is always on, as in C++
This moves the two implementations closer together.  There are still
some divergences in spec/builtin-trap to fix.

I added a hack to restore the default SIGINT handler while running
CPython's raw_input().

This could be changed to a fork of raw_input().  It only affects the dev
build though, so it's not high priority.

In addition to spec tests and spec/stateful, I manually tested that we
fix the "trap INT not running" bug from Samuel in both C++ and Python.