[CommandEvaluator] GC and check for pending traps less often.
I introduced the _LeafTick() function, and it's now called in SOME
branches of _Dispatch().

Prior to this change, we periodic GC/trap work in _Execute(), which is
like calling it at every node in the AST, rather than just the leaves.

This gives a tiny improvement on benchmarks2/gc-cachegrind.  But the I
also think it makes the code clearer, e.g. when tracing when traps are
run.