[spec/ysh-bugs] Repro for issue #2078, pgen2 limit
Also added a unit test.

I tried to make this a bit more efficient in terms of allocations
(before fixing the bug).

Then I realized that PNodeAllocator is a GC object, so it can't allocate
a std::vector in the constructor, because it would never get
freed.

GC objects can't have members that need destruction, like std::vector.