[ysh breaking] Remove eval (cmd) in favor of io->eval(cmd)
As mentioned in the last change, eval $mystr uses a slightly different
algorithm to evaluate code -- main_loop.Batch.  Error handling is
different.

This caused some parsing problems in yblocks.ysh - the "command ends
with expression" problem.

I worked aronud it with with { } and fopen { } wrappers.  I made a note
on Zulip about it.