[test/spec] Refine and test definition of invokable Obj
We should be able to hook it up to core/executor.py with this code.

We get a (me, __invoke__) tuple, and then we can add me=Obj as a keyword
argument.

Do we need a BoundProc?  It doesn't seem like it.

The Obj itself is invokable.  We don't need an (obj, method name) pair
like we do in BoundFounc.