[ysh] Implement invokable Obj with __invoke__ magic method
Documented in

- doc/ref/toc-ysh
- doc/proc-func

See #language-design Zulip for the motivation.  Summary:

It started as a case in spec/ysh-proc-meta, motivated by generating
procs dynamically.

We had solutions based on:

1. eval $mystr
2. parseCommand() io->eval()

Now we have a solution based on:

3. Invokable objects.  The __invoke__ method makes an Obj instance
   "proc-like".

I think this will be big!  For "maximalist YSH".

We can use it for the ctx builtin, which is used by the flag parser.
And Hay, Markaby-style HTML generation, ...