[builtin/use] Basic version of ctx_ModuleEval and caching
It kinda works, but I need to improve the tests.

And ctx_ModuleEval should not use the "front frame".  Because then it
could read from the calling module!  It should just replace the global
frame, which is mem.var_stack[0].

I also need to add the __invoke__ method.  That should be a value.Proc,
although the body will be written in typed Python, not YSH.