[pp test_] Fix bug in cycle detection, and simplify it
Prior to this change, there was an infinite loop bug

I borrowed the logic from display/pp_value.py.

There is still a quirk with Objects that contain a dict with a cycle.
Though I think this is an artifact of printing objects with {}.

I think we might want to print them with <> or something.