[ysh semantics] value.Expr closures in a loop works
Str.replace() still needs some work.

But I removed the mechanism of putting $0 in the var_stack, and then
looking it up there.

Instead, we mutate mem.dollar0.  This is consistent with the way we
handle other vars, and fixes some quirks.

---

Though in the future, it may be nice to deprecate the mem.argv_stack and
mem.dollar0, and put ALL of $0 $1 $2 in the var_stack.  Right now that
is a little tricky, and it's an orthogonal change.