[ysh] Prototype of ENV dict
It's behind shopt --set no_copy_env, so it's not enabled yet

Problem: I don't think we can "silently change what 'export' means.
Because:

    export FOO=bar
    echo $FOO

can still be used.  This would be the new behavior:

    export FOO=bar
    echo $[ENV.FOO]

I suppose that is not too bad, but it may cause problems upon upgrade.

It may be better to disable the 'export' builtin altogether, in favor
of:

    setglobal ENV.FOO = 'bar'