[spec/vars-special] Fix flaky test
osh-native had a spec test delta for case #8.

I suspect it was some variant of the 'stdin has both code and data' problem in shell.

It only happened in the spec test framework, not when running
_bin/cxx-opt/osh interactively.

This is an errata for Oils 0.24.0.  I believe it's a test issue, not a
code issue.