[perf fix] Add manual GC points to 'if' and 'case'
This resolves issue #2123.

The 'filter' loop would evaluate expressions, and allocate, without
every hitting a GC point.

To break out of a loop or recursion, you will need 'if' or 'case'.  So
I believe this should work everywhere, just as it does the
test/bug-2123.* repro.

I'm not quite sure why the bug depended on get() NOT having a default
arg though.  Could look into that more.