[ASDL] Support single arg constructor for variant subtype.
TODO: this might end up as

    CompoundWord.Take(other)  # python

    CompoundWord::Take(other)  // C++

Because I noticed that the List<T>(other) constructor does not obey
invariants.

You can't ALIAS the slab_ field, and also copy the len_ field.  If you
do, then you get a bug:

    strs = ['a', 'b']
    c = CompoundWord(strs)
    strs.append(c)  # modifies slab_ for c, but NOT the len_!

So it appears you have to "clear" the argument!  (or copy it, which we
don't want for efficiency reasons)