[bin] Fix bin name detection
When we have multiple binaries of oil/osh/ysh of different versions,
we want to call these binaries through the names `osh-0.23.0`,
`osh-0.24.0`, etc.  However, with the current implementation of
switching the mode by the basename of the command, one cannot use
those versioned names to call oil/osh/ysh.  This PR loosens the
condition so that a prefix of the basename may match one of the
command names.

Note: This PR would make the C++ version work, but the Python version
still does not work because it tries to find related libraries based
on the path `$0`.