[pretty] Apply ASDL subtype optimization to doc::Concat()
This only reduces memory usage by a little bit, probably because the
doc::Concat() we're saving is a 16-byte object that fits in the small
pool.

But it's nice to know that the optimization works.  And it is easy to
apply, with the aid of MyPy.

TODO: instead of List_Measured.Take([]), we need List_Measured.New().