[refactor] Rename `${#arr[@]}` from `_Length` to `_Count` (#2163)
This is because there are two types of "lengths" in arrays.  One is
the count of elements.  The other is the max index plus one.  They
differ for sparse arrays.  For example, for an array constructed by
`a=(); a[0]=x a[9]=y`, the former is 2, while the latter is 10.  This
patch suggests calling the former "count", and the latter "length".

* [core/state] Fix length in the out-of-bound error message