[mycpp] Implement stable variable naming for global strings
It's not turned on yet, but there is a demo at:

    mycpp/TEST.sh str-hash-demo

This lets us diff two versions of oils_for_unix.mycpp.cc without seeing
"noise"!

Instead of naming the variables

    str1, str2, str3, ...

They are named

    S_zyA, S_pmy, S_ewi, ...

Where the three-letter codes are 15 bit hashes.  Each letter comes from
an alphabet of 32, which is 5 bits.

So 15 bits gives us a space of 32,768 hashes for our ~1580 strings.

Collisions are resolved by adding a number, like S_ewi_1, S_ewi_2, ...