[mycpp refactor] SimpleVisitor handles for loops, list comp
I'm trying to move the computation of local_vars out of the decl pass,
into the virtual pass.

In addition to assignments, loops create local variables:

    for i, item in enumerate(mylist):  # new local i
        pass

And so do list comprehensions:

    result = [x for x in mylist]  # new local x

Note that we only allow list comprehensions on the RHS of an assignment
to a NameExpr.  We don't allow any of:

    myfunc([x for x in mylist])
    a[i] = [x for x in mylist]
    a, b = [x for x in mylist]