[lazylex] Working on <script> <style>
The data inside them is CDATA.  This is actually a lexer mode ...

We just skip everything until </script> </style>.

There can't be more tags between them.  We also allow unescaped < > &,
although we also have to respect &lt;script> too.

I noticed that > is allowed in raw data, but not <.  Hm.