[doctools/micro-syntax] Start HTM8 lexer.
The motivation was to test some HTM8 rules in re2c.

It works fine!  We can distinguish StartEndTag with /> vs. StartTag with
>.

If the matches are equal length, then it picks the first one.

Our rule in lazylex/html.py is different - the first match - but it also
works in this case.