[htm8] Work on AttrLexer
ReadName() and AttrNameEquals() kinda work.

Now I need to work on lexing values.

I discovered a problem with the previous definition of HTM8.  Something
like:

    <img src=?foo=bar/>

is hard to lex because the /> can be part of a quoted attribute, or it
can be a self-closing tag.

Although right now we are just excluding /.  Hm.  We should probably
enforce this for readability:

    <img src="?foo=bar/">