[micro-syntax] Minor HTM8 cleanup
And I realized that these problems are similar:

- shell here doc - search until prefix-selected delimiter (line-wise)
- C++ multi-line string - search until prefix-selected delimiter
- HTML <script> <STYLE> - search until prefix-selected delimiter

And then these are similar:

- C++ /* - search until */
- HTML
  - <!-- search until -->
  - <? search until ?>
  - <![CDATA[ search until ]]>

I'm also not sure if <!DOCTYPE foo> needs to be multi-line.  Probably.

Maybe we need a mode that's not multi-line then?  It will work better
for start tags too.

    <a href=foo
       class=d></a>

Well actually we have already taken care of that, because <a and > are
different tokens.  The mode persists across lines.