print packed completions from minimal display
This commit changes the display style of the minimal completion display.
Instead of only printing a subset of matches in a single column and
saying that there are more, we now just print them all in the packed
format. This behavior will be more similar to what bash users are
familiar with and just generally more useful.