[osh] Fix unquoted `$*`, `${*}`, `${a[*]}`, etc. under empty IFS (#2248)
This is issue #627 (from 2020!)

In the current master, even when multiple elements are set to $*,
unquoted $* produces a single connected word if IFS is an empty
string.  However, we should not join the elements for unquoted $* even
if IFS is an empty string.

This is related to how $* and $@ are processed internally.  We have
been first joining $* and $@ with the first character of IFS and later
split it again with IFS.  We kept the array structure only for the
"$@" case.  However, this way, when IFS is an empty string, separate
elements in $* and $@ are unexpectedly joined and cannot be split
again.

In this patch, we shall keep the array structure also for $* and $@ in
addition to "$@".  However, we also need to carry the information to
distinguish $* or $@ from "$@".  For this purpose, we add a new member
"quoted" to part_value.Array.

---

This change appears to change behavior in the case of:

- IFS=x (as opposed to default or empty )
- $@ - unquoted
- empty string like set -- one '' two
  - but NOT set -- 'a b' c ''

We made note of those in these spec test files:

- spec/toysh-posix - Clarify affected test cases
- spec/word-split - Enhance test cases

---------

Co-authored-by: Andy C <andy@oilshell.org>