[test/ble.sh] Add "--lib" in loading "out/ble.osh"
The "osh" branch in the upstream akinomyoga/ble.sh has been rebased on
the latest version of ble.sh.  As a consequence, sourcing out/ble.osh
in the CI has started to fail.  The latest version of ble.sh checks
whether the current session has an associated TTY when loaded as an
interactive setting.  When ble.sh cannot detect the TTY, it cancels
its loading.  To load ble.sh in a non-interactive session, one needs
to source ble.sh using ``the library mode'' by specifying the option
"--lib".