[osh] Add temporary adjustment for "assoc=(1 2)"
The current OSH behavior of overwriting "assoc" with a new indexed
array for "assoc=(1 2)" conflicts with the Bash 5.1 feature
"assoc=(key value)".  The current master behavior will soon be changed
when the Bash 5.1 feature is implemented, yet we keep the current
master behavior for now to avoid changing the behavior too many times.
This commit adjusts the behavior to be backward compatible with the
current master version.