[spec/ysh-user-feedback] Failing test cases for feedback related to static var and setvar
From Julian, Aidan, Machine Stops

OK we should get rid of those static checks.

[demo] Test out the equivalent in JavaScript

Funny thing: I found a corner of JavaScript where you can either mutate
or redeclare the loop variable, with 'let'.  I thought that you would
not be allowed to redeclare it - because there should be a single
block/scope for the loop body?

But it appears there are two?

More JavaScript oddities.