[test/spec] Failing test cases for ble.sh
Mentioned in issue #1069.

[spec/ble-idioms] Failing test case for array=99 and assoc=99

[spec/builtin-meta] Investigating the \builtin local issue

The problem is that we do detection of

    builtin local a=$x
    builtin builtin local a=$x

to deteect the special assignment builtin evaluation rules.

But we only look at one "part value", and

    \builtin

evaluates to 2 part values, for 'b' and builtin.

Solutions:

- Look at arbitrary numbers of part values
- Don't use this dynamic algorithm

Instead, we can accept that we don't detect

    \builtin
    b=builtin

In those cases, we can evaluate to a regular cmd_value.Argv.  And then
do word splitting anyway.

In the dynamic case of

    b=builtin

No shell does the proper detection.