[osh] Turn on rule for special builtins and prefix bindings
It's not hidden behind set -o posix

For example, eval set : special builtins, and prefix bindings are
persist in the current stack frame.

They are NOT "temp bindings", which means that they push a new stack
frame.

This makes this ble-idioms test case work:

    f() {
      IFS= eval 'local x=x'
      echo $x
    }

Both 'eval' and 'local' are special builtins, so the binding of x
persists afterward.

This is part of issue #1069, for ble.sh.