[test/spec] Introduce OK-2, OK-3, OK-4 assertions
With different colors

Use these assertions in spec/ble-unset, because there are many different
behaviors.  This lets us see the patterns of behavior.

Right now we're using OK, OK-2, and OK-3.

But not OK-4.

Combined with PASS, there are cases with 4 different behaviors.  It's a
bit arbitrary whether we consider any of them BUG.