[doctools] Start rewriting the cmark build; remove cmake dependency
I realized that the shared library is hard to build with a shell script,
and we should use a subprocess instead.  It won't cost any time.

I tested that bin/cmark behaves the same as the shared library, but it's
not turned on yet.  That will be next.

Related: speed up 'build/py.sh all' by building doc/ref/* in parallel.