[core] Implement POSIX rule for $PWD
This fixes the Perl build failure.

If $PWD is an absolute path, and it's the same file as '.', then use it
as our "cached" $PWD value.

This means that symlinks to $PWD work.

Otherwise, recompute it.

Thanks to Samuel for finding this bug, and narrowing it down.

It still took some effort to find on top:

- grepping through dash source code (but missing this important part)
- uploading dash source to Claude and asking questions
- finding the reference to the POSIX rule in bash source code